Move migrations and design time context around added console app that can be used with ef tools to deploy db migrations
