+++ RecipeCrawler.DatabaseMigrator/RecipeCrawler.DatabaseMigrator.csproj
@@ -27,4 +27,8 @@
     <ProjectReference Include="..\RecipeCrawler.Data\RecipeCrawler.Data.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Migrations" />
+  </ItemGroup>
+
 </Project>
